+++ CHANGELOG.md
@@ -4,16 +4,28 @@
 
 * AADAccessReviewDefinition
   * Initial release.
+* AADAuthenticationMethodPolicyExternal
+  * Initial release.
 * AADCustomSecurityAttributeDefinition
   * Fixed missing permissions in settings.json
 * AADIdentityB2XUserFlow
   * Initial release.
 * AADIdentityGovernanceProgram
   * Initial release.
+* AADIdentityAPIConnector
+  * Initial release.
+* AADOrganizationCertificateBasedAuthConfiguration
+  * Initial release.
 * AADSocialIdentityProvider
   * Fixed missing permissions in settings.json
+* EXOMailboxAuditBypassAssociation
+  * Initial release.
 * Intune workload
   * Fixed missing permissions in settings.json
+* SentinelAlertRule
+  * Initial release.
+* SentinelThreatIntelligenceIndicator
+  * Initial release.
 * SCDeviceTenantPolicy
   * Initial Release.
 * SPOTenantSettings
@@ -46,6 +58,8 @@
   * Initial release.
 * AADConnectorGroupApplicationProxy
   * Initial release.
+* AADCustomAuthenticationExtension
+  * Initial release.
 * AADCustomSecurityAttributeDefinition
   * Initial release.
 * AADDeviceRegistrationPolicy
@@ -57,7 +71,8 @@
 * AADLifecycleWorkflowSettings
   * Initial release.
 * AADServicePrincipal
-  * Adding Delegated Permission Classification Property
+  * Added Delegated Permission Classification Property
+  * Added Custom Security Attributes Property
 * ADOPermissionGroupSettings
   * Initial release.
 * EXOATPBuiltInProtectionRule